requirements.txt
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0
google-generativeai==0.5.4
schedule==1.2.0
pync==2.0.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0